chore: Main
main.dart 에서 calculator 페이지 분리
